requirements.txt
absl-py==1.4.0
aioice==0.9.0
aiortc==1.5.0
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.4.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
av==10.0.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.2
cmd2==2.4.3
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.0
cryptography @ file:///croot/cryptography_1677533068310/work
cuda-python==12.2.0
cycler==0.11.0
Cython==3.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
dnspython==2.3.0
entrypoints==0.4
etils==1.4.1
exceptiongroup==1.1.1
executing==1.2.0
fake-useragent==1.1.3
fastjsonschema==2.17.1
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.0
fqdn==1.5.1
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645438755360/work
google-auth==2.19.1
google-auth-oauthlib==1.0.0
google-crc32c==1.5.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
humanfriendly==10.0
idna==3.4
ifaddr==0.2.0
imageio==2.31.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
imread-from-url==0.1.3
install==1.3.5
ipykernel==6.25.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jax==0.4.11
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
lit==16.0.6
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
MechanicalSoup==0.8.0
mistune==2.0.5
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
MyNeuron==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.5.0
nbformat==5.9.0
nest-asyncio==1.5.7
networkx==3.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.2
nvidia-cublas-cu11==2022.4.8
nvidia-cublas-cu117==11.10.1.25
nvidia-cublas-cu12==12.2.4.5
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.2.128
nvidia-cuda-runtime-cu12==12.2.128
nvidia-cudnn-cu11==8.6.0.163
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvidia-pyindex==1.0.9
nvidia-tensorrt==99.0.0
oauthlib==3.2.2
onnx==1.12.0
onnx-graphsurgeon==0.3.27
onnx2tf==1.13.12
onnxruntime==1.12.1
onnxruntime-gpu==1.15.1
onnxsim==0.4.33
opencv-python==4.8.0.74
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
outcome==1.2.0
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
patool==1.12
pbr==5.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==3.10.0
prettytable==0.7.2
progressbar2==3.34.3
prometheus-client==0.17.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.10.4
pycparser==2.21
pycuda==2022.2.2
pydeck==0.8.0
pyee==9.1.0
Pygments==2.15.1
pylibsrtp==0.8.0
Pympler==1.0.1
pyngrok==6.0.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyperclip==1.8.2
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-utils==3.5.2
pytools==2023.1.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
roboflow==1.0.9
rpds-py==0.9.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
selenium==4.10.0
Send2Trash==1.8.2
sentry-sdk==1.23.1
six==1.16.0
smmap==5.0.0
sng4onnx==1.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.4.1
stack-data==0.6.2
stevedore==5.1.0
streamlit==1.22.0
streamlit-drawable-canvas==0.9.3
streamlit-webrtc==0.45.0
supervision==0.12.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-cpu==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
tensorrt==8.6.0
tensorrt-bindings==8.6.1
tensorrt-dispatch==8.6.1
tensorrt-dispatch-bindings==8.6.1
tensorrt-dispatch-libs==8.6.1
tensorrt-lean==8.6.1
tensorrt-lean-bindings==8.6.1
tensorrt-lean-libs==8.6.1
tensorrt-libs==8.6.1
termcolor==2.3.0
terminado==0.17.1
tf2onnx==1.8.1
tflite-support==0.4.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.7.18
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-tensorrt==1.4.0
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
triton==2.0.0
typing_extensions @ file:///croot/typing_extensions_1681939499988/work
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.0.147
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1680254681959/work
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.6
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wget==3.2
widgetsnbextension==4.0.8
wrapt==1.14.1
wsproto==1.2.0
zipp==3.15.0
